+++ ChangeLog.md
@@ -7,6 +7,7 @@
 ### Changed
 - 连接配置，回调函数的参数可选
 - `PDOFactory`重命名为`PDOProvider`
+- 变更`PDOProviderInterface`接口
 
 ### Removed